# docs/swagger.yaml
basePath: /api
definitions:
  common.APIResponse:
    description: 所有 API 响应的统一包装格式，包含状态码、消息、数据和 timestamp
    properties:
      code:
        description: 状态码，200 表示成功，其他表示错误
        example: 200
        type: integer
      data:
        description: 响应数据，成功时返回业务数据，错误时为 null
      message:
        description: 消息提示，成功时为"success"，错误时为错误描述
        example: success
        type: string
      timestamp:
        description: Unix 时间戳（秒）
        example: 1717200000
        type: integer
    type: object
...
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
  models.Image:
    description: 系统上传的图片资源信息
    properties:
      category:
        description: 分类
        example: banner
        type: string
      createdAt:
        description: 创建时间
        example: "2024-01-01T00:00:00Z"
        type: string
      deletedAt:
        allOf:
        - $ref: '#/definitions/gorm.DeletedAt'
        description: 删除时间（软删除）
      description:
        description: 图片描述
        example: 首页横幅图片
        type: string
      filePath:
        description: 文件路径
        example: /uploads/images/banner.jpg
        type: string
      fileSize:
        description: 文件大小（字节）
        example: 102400
        type: integer
      filename:
        description: 文件名
        example: banner.jpg
        type: string
      id:
        description: 图片 ID
        example: 1
        type: integer
      longDescription:
        description: 详细描述
        example: 这是首页展示的横幅图片
        type: string
      sortOrder:
        description: 排序顺序
        example: 1
        type: integer
      updatedAt:
        description: 更新时间
        example: "2024-01-01T00:00:00Z"
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: |-
    医疗器械内容管理系统 API 接口文档
    提供内容管理、模块配置、图片管理、多语言支持等功能
  title: Shirley CMS API
  version: "1.0"
paths:
...
  /api/admin/images:
    get:
      consumes:
      - application/json
      description: 获取所有或指定分类的图片列表
      parameters:
      - description: 图片分类，不传则返回所有图片
        in: query
        name: category
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 图片列表
          schema:
            allOf:
            - $ref: '#/definitions/common.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Image'
                  type: array
              type: object
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/common.APIResponse'
      security:
      - APIKey: []
      summary: 获取图片列表
      tags:
      - 图片管理
    post:
      consumes:
      - multipart/form-data
      description: 上传单张图片到服务器并记录到数据库
      parameters:
      - description: 图片文件
        in: formData
        name: image
        required: true
        type: file
      - description: 图片描述
        in: formData
        name: description
        type: string
      - description: 图片分类
        in: formData
        name: category
        type: string
      - description: 排序顺序
        in: formData
        name: sort_order
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 上传成功
          schema:
            allOf:
            - $ref: '#/definitions/common.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Image'
              type: object
        "400":
          description: 未上传文件
          schema:
            $ref: '#/definitions/common.APIResponse'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/common.APIResponse'
      security:
      - APIKey: []
      summary: 上传单张图片
      tags:
      - 图片管理
  /api/admin/images/{id}:
    delete:
      consumes:
      - application/json
      description: 根据图片 ID 删除图片
      parameters:
      - description: 图片 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/common.APIResponse'
        "404":
          description: 图片不存在
          schema:
            $ref: '#/definitions/common.APIResponse'
      security:
      - APIKey: []
      summary: 删除图片
      tags:
      - 图片管理
    put:
      consumes:
      - multipart/form-data
      description: 更新指定图片的描述、分类等信息，支持重新上传图片
      parameters:
      - description: 图片 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 新的图片文件
        in: formData
        name: image
        type: file
      - description: 图片描述
        in: formData
        name: description
        type: string
      - description: 图片详细描述
        in: formData
        name: long_description
        type: string
      - description: 图片分类
        in: formData
        name: category
        type: string
      - description: 排序顺序
        in: formData
        name: sort_order
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            allOf:
            - $ref: '#/definitions/common.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Image'
              type: object
        "404":
          description: 图片不存在
          schema:
            $ref: '#/definitions/common.APIResponse'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/common.APIResponse'
      security:
      - APIKey: []
      summary: 更新图片信息
      tags:
      - 图片管理
  /api/admin/images/batch:
    post:
      consumes:
      - multipart/form-data
      description: 批量上传多张图片到服务器
      parameters:
      - description: 图片文件数组
        in: formData
        name: images
        required: true
        type: file
      - description: 图片分类
        in: formData
        name: category
        type: string
      - collectionFormat: csv
        description: 图片描述数组
        in: formData
        items:
          type: string
        name: descriptions
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: 上传成功
          schema:
            allOf:
            - $ref: '#/definitions/common.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Image'
                  type: array
              type: object
        "400":
          description: 未上传文件
          schema:
            $ref: '#/definitions/common.APIResponse'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/common.APIResponse'
      security:
      - APIKey: []
      summary: 上传多张图片
      tags:
      - 图片管理
  /api/admin/images/by-filename/{filename}:
    delete:
      consumes:
      - application/json
      description: 根据图片文件名删除图片
      parameters:
      - description: 图片文件名
        in: path
        name: filename
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/common.APIResponse'
        "404":
          description: 图片不存在
          schema:
            $ref: '#/definitions/common.APIResponse'
      security:
      - APIKey: []
      summary: 根据文件名删除图片
      tags:
      - 图片管理
...
securityDefinitions:
  APIKey:
    description: JWT Token 认证，格式：Bearer {token}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"